Move root files to Root routes directory
