+++ src/configureStore.js
@@ -2,18 +2,17 @@ import { applyMiddleware, createStore, compose } from 'redux'
 import { composeWithDevTools } from 'redux-devtools-extension/logOnlyInProduction'
 import createSagaMiddleware from 'redux-saga'
 
-// @todo Move to other directories
-import rootReducer from './rootReducer'
-import rootSaga from './rootSaga'
+import rootReducer from 'routes/Root/rootReducer'
+import rootSaga from 'routes/Root/rootSaga'
 
 const sagaMiddleware = createSagaMiddleware()
 
 const middlewares = [
-  sagaMiddleware,
+  sagaMiddleware
 ]
 
 const enhancers = [
-  composeWithDevTools(applyMiddleware(...middlewares)),
+  composeWithDevTools(applyMiddleware(...middlewares))
 ]
 
 const composedEnhancer = compose(...enhancers)
@@ -24,8 +23,7 @@ const composedEnhancer = compose(...enhancers)
  * @return { object } store - store
  * @todo Check environment and only execute debugEnhancer if in development
  */
-export default function configureStore(initialState = {}) {
-
+export default function configureStore (initialState = {}) {
   const store = createStore(
     rootReducer,
     initialState,
