+++ src/main.js
@@ -6,8 +6,8 @@ import { Router, browserHistory } from 'react-router'
 import { syncHistoryWithStore } from 'react-router-redux'
 
 import App from 'components/App'
-import configureStore from "./configureStore"
-import rootRoute from "./rootRoute"
+import configureStore from './configureStore'
+import rootRoute from 'routes/Root/rootRoute'
 
 const store = configureStore()
 const history = syncHistoryWithStore(browserHistory, store)
@@ -24,7 +24,6 @@ ReactDOM.render(
   ,
   document.getElementById('root')
 )
-
 
 // if (module.hot) {
 //   module.hot.accept('./components/App', () => {
